requirements.txt
undetected-chromedriver
selenium
beautifulsoup4
telethon
fastapi
python-dotenv
requests
uvicorn
pandas
python-telegram-bot[job-queue]
